+++ package.json
@@ -4,7 +4,9 @@
   "description": "A simple visual demo using html5 canvas and javascript",
   "main": "fire.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc && npm run copy-assets",
+    "copy-assets": "cp -R src/assets dist/assets && cp -R src/styles dist/styles && cp src/index.html dist && cp src/favicon.ico dist"
   },
   "repository": {
     "type": "git",
