+++ src/index.html
@@ -5,13 +5,13 @@
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Canvas Fire Animation</title>
-	<link rel="stylesheet" href="style.css">
+	<link rel="stylesheet" href="styles/style.css">
 	<link rel="icon" href="favicon.ico" type="image/x-icon">
 </head>
 
 <body>
 	<canvas id="fireCanvas"></canvas>
-	<script src="fire.js"></script>
+	<script src="scripts/fire.js"></script>
 </body>
 
 </html>
