+++ src/scripts/fire.ts
@@ -1,6 +1,16 @@
-document.addEventListener("DOMContentLoaded", function () {
-	const canvas = document.getElementById('fireCanvas');
+import { hslToRgb, rgbToHsl } from './rgbhsl';
+
+function initializeFireCanvas() {
+	const canvas = document.getElementById('fireCanvas') as HTMLCanvasElement;
+	if (!canvas) {
+		console.error("Canvas not found!");
+		return;
+	}
 	const ctx = canvas.getContext('2d');
+	if (!ctx) {
+		console.error("Canvas context not found!");
+		return;
+	}
 
 	const lowResWidth = 320;
 	const lowResHeight = 200;
@@ -19,52 +29,6 @@ document.addEventListener("DOMContentLoaded", function () {
 	let holdX = 0;
 	let holdY = 0;
 
-	function hslToRgb(h, s, l) {
-		s /= 100;
-		l /= 100;
-		let c = (1 - Math.abs(2 * l - 1)) * s;
-		let x = c * (1 - Math.abs((h / 60) % 2 - 1));
-		let m = l - c / 2;
-		let r = 0, g = 0, b = 0;
-
-		if (0 <= h && h < 60) { r = c; g = x; b = 0; }
-		else if (60 <= h && h < 120) { r = x; g = c; b = 0; }
-		else if (120 <= h && h < 180) { r = 0; g = c; b = x; }
-		else if (180 <= h && h < 240) { r = 0; g = x; b = c; }
-		else if (240 <= h && h < 300) { r = x; g = 0; b = c; }
-		else if (300 <= h && h < 360) { r = c; g = 0; b = x; }
-
-		r = Math.round((r + m) * 255);
-		g = Math.round((g + m) * 255);
-		b = Math.round((b + m) * 255);
-
-		return [r, g, b];
-	}
-
-	function rgbToHsl(r, g, b) {
-		r /= 255;
-		g /= 255;
-		b /= 255;
-		let max = Math.max(r, g, b);
-		let min = Math.min(r, g, b);
-		let h, s, l = (max + min) / 2;
-
-		if (max === min) {
-			h = s = 0;
-		} else {
-			let d = max - min;
-			s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
-			switch (max) {
-				case r: h = (g - b) / d + (g < b ? 6 : 0); break;
-				case g: h = (b - r) / d + 2; break;
-				case b: h = (r - g) / d + 4; break;
-			}
-			h *= 60;
-		}
-
-		return [Math.round(h), s * 100, l * 100];
-	}
-
 	const reseed = () => {
 		for (let x = 0; x < lowResWidth; x++) {
 			let index = (x * 4) + 0 * (lowResWidth * 4);
@@ -82,7 +46,7 @@ document.addEventListener("DOMContentLoaded", function () {
 		hueShift = (hueShift + 1) % 360;
 	}
 
-	function updateCanvas(timestamp) {
+	function updateCanvas(timestamp: number) {
 		if (timestamp - lastTime > timingInterval) {
 			for (let y = 1; y < lowResHeight; y++) {
 				for (let x = 0; x < lowResWidth; x++) {
@@ -95,9 +59,9 @@ document.addEventListener("DOMContentLoaded", function () {
 					heatMap[y * lowResWidth + x] *= 0.3;
 					let heat = heatMap[y * lowResWidth + x];
 					let heatintensity = Math.min(heat / 5, 20);
-					r = Math.max(0, (pixels[aboveIndex] * factor + pixels[index]) / (factor + .92) - 1) + heatintensity;
-					g = Math.max(0, (pixels[aboveIndex + 1] * factor + pixels[index + 1]) / (factor + .92) - 1) + heatintensity;
-					b = Math.max(0, (pixels[aboveIndex + 2] * factor + pixels[index + 2]) / (factor + .92) - 1) + heatintensity;
+					let r = Math.max(0, (pixels[aboveIndex] * factor + pixels[index]) / (factor + .92) - 1) + heatintensity;
+					let g = Math.max(0, (pixels[aboveIndex + 1] * factor + pixels[index + 1]) / (factor + .92) - 1) + heatintensity;
+					let b = Math.max(0, (pixels[aboveIndex + 2] * factor + pixels[index + 2]) / (factor + .92) - 1) + heatintensity;
 
 					if (heat > 0) {
 						let [h, s, l] = rgbToHsl(r, g, b);
@@ -118,14 +82,14 @@ document.addEventListener("DOMContentLoaded", function () {
 				applyHeat(holdX, holdY);
 			}
 
-			ctx.putImageData(imageData, 0, 0);
+			ctx?.putImageData(imageData, 0, 0);
 			lastTime = timestamp;
 			reseed();
 		}
 		requestAnimationFrame(updateCanvas);
 	}
 
-	function applyHeat(x, y) {
+	function applyHeat(x: number, y: number) {
 		let rect = canvas.getBoundingClientRect();
 
 		let canvasX = Math.floor((x / rect.width) * lowResWidth);
@@ -148,7 +112,7 @@ document.addEventListener("DOMContentLoaded", function () {
 		}
 	}
 
-	function startHold(event) {
+	function startHold(event: MouseEvent | Touch) {
 		isHolding = true;
 		holdX = event.clientX;
 		holdY = event.clientY;
@@ -158,7 +122,7 @@ document.addEventListener("DOMContentLoaded", function () {
 		isHolding = false;
 	}
 
-	function updateHoldPosition(event) {
+	function updateHoldPosition(event: MouseEvent | Touch) {
 		if (isHolding) {
 			holdX = event.clientX;
 			holdY = event.clientY;
@@ -184,4 +148,6 @@ document.addEventListener("DOMContentLoaded", function () {
 
 	reseed();
 	requestAnimationFrame(updateCanvas);
-});
+}
+
+document.addEventListener("DOMContentLoaded", initializeFireCanvas);
